Modified start script

+++ start.sh
@@ -1,16 +1,21 @@
 #!/bin/ash
-echo Please start comment this section to prevent settings overwrite
-echo and change file permissions to 555
-
 git clone https://github.com/SahrulGamerz/ptero-eggs.git ./temp
-cp -r ./temp/nginx ./
-cp -r ./temp/php-fpm ./
-cp -r ./temp/webroot ./
+if [[ -f "./nginx" ]]
+    echo "Cloning nginx folder"
+    cp -r ./temp/nginx ./
+else if [[ -f "./php-fpm" ]]
+    echo "Cloning php-fpm folder"
+    cp -r ./temp/php-fpm ./
+else if [[ -f "./webroot" ]]
+    echo "Cloning webroot folder"
+    cp -r ./temp/webroot ./
+then
+    echo "File exists not cloning files"
+fi
+
 cp ./temp/start.sh ./
 chmod +x ./start.sh
 rm -rf ./temp
-
-echo End of comment
 
 echo "Starting PHP-FPM..."
 php-fpm8 --fpm-config /home/container/php-fpm/php-fpm.conf --daemonize
